+++ stories/components/ellipsis/docs/examples/basic_usage.tsx
@@ -1,8 +1,9 @@
 import { Box } from '@semcore/base-components';
+import type { EllipsisProps } from '@semcore/ellipsis';
 import Link from '@semcore/link';
 import React from 'react';
 
-const Demo = () => {
+const Demo = (props: EllipsisProps) => {
   return (
     <Box w={220}>
       <Link w={220} href='https://developer.semrush.com/intergalactic/components/ellipsis/ellipsis'>
@@ -13,5 +14,13 @@ const Demo = () => {
     </Box>
   );
 };
+
+export const defaultProps: EllipsisProps = {
+  trim: 'end',
+  tooltip: true,
+  maxLine: undefined,
+};
+
+Demo.defaultProps = defaultProps;
 
 export default Demo;