+++ semcore/ellipsis/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.6] - 2025-08-08
+
+### Fixed
+
+- Ellipsis with `trim=middle` cut off text at the end.
+- Incorrect cropping with different font-size.
+
 ## [16.0.5] - 2025-07-23
 
 ### Changed